+++ src/internal/adapter/web/FullscreenAPI.ts
